Switch from actors to message passing with channels
